// libs/eventstore-interconnect/src/module/legacy/event-store-config.interface.ts
import { UserCredentials } from 'geteventstore-promise';
import { ConnectionSettings, TcpEndPoint } from 'node-eventstore-client';
import { EventStore } from './event-store';

export type IHttpEndpoint = TcpEndPoint;

export interface IEventStoreConfig {
  credentials: UserCredentials;
  tcpConnectionName?: string;
  tcp?: TcpEndPoint;
  http?: IHttpEndpoint;
  clusterDns?: string;
  options?: ConnectionSettings;
  onTcpConnected?: (eventStore: EventStore) => void;
  onTcpDisconnected?: (eventStore: EventStore) => void;
}